fix: moved the textcontent update to fix jump
On ties and loss the text box was being updated to a length that was pushing the other two pieces aside before the animation started, made an ugly jump. Moving the textContent update to before the timeout pause so it happens during the transition of gameboard to resultsboard and essentially hides the jump.